Trust QR fix round 1: entitlement tests, welcome.rs agent identity
Completes the partial fix left mid-edit in the working tree (findings 1-3's
parsing was already done; added the missing entitlements test coverage) and
implements finding 4 from scratch: welcome.rs's authorization QR now uses
the agent keypair slot and DEFAULT_AGENT_ENTITLEMENTS instead of the device
keypair and hardcoded agent.capability.* strings, matching agent.rs's
--trust flow.

Also restores parse_agents_config's public visibility in agent.rs
(pub(crate) use -> pub use), which Task 2's agent_config.rs extraction had
silently narrowed, breaking three integration tests that import it via
arkavo_cli::commands::agent::parse_agents_config. Discovered by running the
full cargo test -p arkavo-cli command instead of --lib only.

Registers the untracked agent_config_parse.rs module (missing mod
declaration left the tree uncompilable). Splits two module doc comments to
clear clippy's too_long_first_doc_paragraph lint under -D warnings.
